# openapi/core/api.yaml
openapi: 3.0.3
info:
  title: Team Api
  description: Team HTTP APIs definition.
  version: 0.1.2
  contact:
    email: smokybeans@zextras.com
servers:
  - url: https://zextras.com/zx/team
...
paths:
...
  /attachments/{fileId}:
    delete:
      tags:
        - Attachments
      summary: Deletes an uploaded attachment
      operationId: deleteAttachment
      parameters:
        - in: path
          name: fileId
          description: file identifier
          schema:
            type: string
            format: uuid
          required: true
          allowEmptyValue: false
      responses:
        204:
          $ref: '#/components/responses/DeleteAttachment204'
        403:
          $ref: '#/components/responses/Forbidden'
    get:
      tags:
        - Attachments
      summary: Retrieves info related to an uploaded attachment
      operationId: getAttachmentInfo
      parameters:
        - in: path
          name: fileId
          description: file identifier
          schema:
            type: string
            format: uuid
          required: true
          allowEmptyValue: false
      responses:
        200:
          $ref: '#/components/responses/GetAttachmentInfo200'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
  /attachments/{fileId}/download:
    get:
      tags:
        - Attachments
      summary: Retrieves an uploaded attachment
      operationId: getAttachment
      parameters:
        - in: path
          name: fileId
          description: file identifier
          schema:
            type: string
            format: uuid
          required: true
          allowEmptyValue: false
      responses:
        200:
          $ref: '#/components/responses/GetAttachment200'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
  /attachments/{fileId}/preview:
    get:
      tags:
        - Attachments
      summary: Retrieves the prefiew of an uploaded attachment
      operationId: getAttachmentPreview
      parameters:
        - in: path
          name: fileId
          description: file identifier
          schema:
            type: string
            format: uuid
          required: true
          allowEmptyValue: false
      responses:
        200:
          $ref: '#/components/responses/GetAttachmentPreview200'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
components:
  responses:
...
    DeleteAttachment204:
      description: The file was deleted correctly
    GetAttachmentInfo200:
      description: Attachment informations
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Attachment'
    GetAttachment200:
      description: The requested file
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary
    GetAttachmentPreview200:
      description: The requested file preview
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary
    NotFound:
      description: The requested resource was not found
    Forbidden:
      description: The requester could not access the resource
...
  schemas:
...
    Attachment:
      type: object
      description: Attachment of a message
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          description: identifier
        name:
          type: string
          description: file name
        size:
          type: integer
          format: int32
          readOnly: true
          description: file length
        mimeType:
          type: string
          readOnly: true
          description: mime type
        userId:
          type: string
          format: uuid
          readOnly: true
          description: identifier of updated user
        roomId:
          type: string
          format: uuid
          readOnly: true
          description: identifier of destination room
      required: [ id, name, size, mimeType, userId, roomId ]